back finalizado

+++ BackSRC/AlmoxarifadoBackSLN/AlmoxarifadoBackAPI/Models/Produto.cs
@@ -7,7 +7,7 @@
         public string Descricao { get; set; }
 
         public string UnMedida { get; set; }
-        public float EstoqueAtual { get; set; }
+        public int EstoqueAtual { get; set; }
         public bool EPermanente { get; set; }
         public int CodigoCategoria { get; set; }
     }
